Renamed internals
