+++ src/main/internal/arrays.ts
