+++ src/main/internal/bitmasks.ts
