+++ src/main/internal/lang.ts
@@ -1,7 +1,5 @@
 export const isArray = Array.isArray;
 
-export const trunc = Math.trunc;
-
 export function identity<T>(value: T): T {
   return value;
 }
