+++ src/main/internal/objects.ts
