+++ src/main/internal/shapes.ts
