+++ src/main/shape/ArrayShape.ts
@@ -1,7 +1,4 @@
 import { CODE_TUPLE, CODE_TYPE, MESSAGE_ARRAY_TYPE, MESSAGE_TUPLE } from '../constants';
-import { createIssueFactory } from '../helpers';
-import { TYPE_ARRAY, TYPE_OBJECT, TYPE_UNKNOWN } from '../Type';
-import { ApplyOptions, ConstraintOptions, Issue, Message } from '../types';
 import {
   applyShape,
   concatIssues,
@@ -14,7 +11,10 @@ import {
   toArrayIndex,
   toDeepPartialShape,
   unshiftIssuesPath,
-} from '../utils';
+} from '../internal';
+import { TYPE_ARRAY, TYPE_OBJECT, TYPE_UNKNOWN } from '../Type';
+import { ApplyOptions, ConstraintOptions, Issue, Message } from '../types';
+import { createIssueFactory } from '../utils';
 import { CoercibleShape } from './CoercibleShape';
 import { AnyShape, DeepPartialProtocol, INPUT, NEVER, OptionalDeepPartialShape, OUTPUT, Result } from './Shape';
