+++ src/main/shape/MapShape.ts
@@ -1,7 +1,4 @@
 import { CODE_TYPE, MESSAGE_MAP_TYPE } from '../constants';
-import { createIssueFactory } from '../helpers';
-import { TYPE_ARRAY, TYPE_MAP, TYPE_OBJECT } from '../Type';
-import { ApplyOptions, ConstraintOptions, Issue, Message } from '../types';
 import {
   applyShape,
   concatIssues,
@@ -14,7 +11,10 @@ import {
   ok,
   toDeepPartialShape,
   unshiftIssuesPath,
-} from '../utils';
+} from '../internal';
+import { TYPE_ARRAY, TYPE_MAP, TYPE_OBJECT } from '../Type';
+import { ApplyOptions, ConstraintOptions, Issue, Message } from '../types';
+import { createIssueFactory } from '../utils';
 import { CoercibleShape } from './CoercibleShape';
 import {
   AnyShape,
