+++ src/main/shape/NumberShape.ts
@@ -6,10 +6,10 @@ import {
   MESSAGE_NUMBER_INTEGER,
   MESSAGE_NUMBER_TYPE,
 } from '../constants';
-import { createIssueFactory } from '../helpers';
+import { getCanonicalValueOf, isArray, isNumber, ok } from '../internal';
 import { TYPE_ARRAY, TYPE_BOOLEAN, TYPE_DATE, TYPE_NUMBER, TYPE_OBJECT, TYPE_STRING } from '../Type';
 import { ApplyOptions, ConstraintOptions, Literal, Message } from '../types';
-import { getCanonicalValueOf, isArray, isNumber, ok } from '../utils';
+import { createIssueFactory } from '../utils';
 import { CoercibleShape } from './CoercibleShape';
 import { AllowLiteralShape, NEVER, ReplaceLiteralShape, Result } from './Shape';
