+++ src/main/shape/PromiseShape.ts
@@ -1,8 +1,8 @@
 import { CODE_TYPE, ERROR_REQUIRES_ASYNC, MESSAGE_PROMISE_TYPE } from '../constants';
-import { createIssueFactory } from '../helpers';
+import { applyShape, copyUnsafeChecks, isArray, ok, toDeepPartialShape } from '../internal';
 import { TYPE_PROMISE } from '../Type';
 import { ApplyOptions, ConstraintOptions, Message } from '../types';
-import { applyShape, copyUnsafeChecks, isArray, ok, toDeepPartialShape } from '../utils';
+import { createIssueFactory } from '../utils';
 import { CoercibleShape } from './CoercibleShape';
 import { AnyShape, DeepPartialProtocol, INPUT, OptionalDeepPartialShape, OUTPUT, Result } from './Shape';
