+++ src/test/internal/types.test.ts
@@ -1,5 +1,5 @@
+import { distributeTypes, unionTypes } from '../../main/internal';
 import { TYPE_NUMBER, TYPE_STRING, TYPE_UNKNOWN } from '../../main/Type';
-import { distributeTypes, unionTypes } from '../../main/utils';
 
 test('unionTypes', () => {
   expect(unionTypes([])).toEqual([]);
