+++ src/test/shapes/SetShape.test.ts
@@ -1,7 +1,7 @@
 import { ApplyOptions, ObjectShape, Ok, Result, SetShape, Shape, StringShape } from '../../main';
 import { CODE_TYPE, MESSAGE_SET_TYPE, MESSAGE_STRING_TYPE } from '../../main/constants';
+import { nextNonce } from '../../main/internal';
 import { TYPE_ARRAY, TYPE_OBJECT, TYPE_SET, TYPE_STRING } from '../../main/Type';
-import { nextNonce } from '../../main/utils';
 
 describe('SetShape', () => {
   class AsyncShape extends Shape {
